+++ src/PepperModSystem.cs
@@ -12,6 +12,9 @@ namespace PepperMod
             api.RegisterBlockEntityClass("PerennialPepperPlant", typeof(BlockEntityPerennialPepperPlant));
             api.RegisterItemClass("ItemPepperSeeds", typeof(ItemPepperSeeds));
             api.RegisterItemClass("ItemPepperFood", typeof(ItemPepperFood));
+            api.RegisterItemClass("ItemPepperBundle", typeof(ItemPepperBundle));
+            api.RegisterBlockClass("BlockHangingPepperBundle", typeof(BlockHangingPepperBundle));
+            api.RegisterBlockEntityClass("HangingPepperBundle", typeof(BlockEntityHangingPepperBundle));
         }
     }
 }